refactor: PaymentAmount 원시값 포장
